requirements.txt
Gunicorn
flask
google.generativeai
pymysql
cryptography